backend/requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
Django==5.1.7
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Flask==3.1.0
geos==0.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==3.0.2
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
Werkzeug==3.1.3
